backend/requirement.txt
pipenv
djangorestframework
django-cors-headers